Poprawki zwiazane z funkcjonalnością (dodano powiadomienia)
-dodano powiadomienia pojawiajace sie przy sukcesie dodania,
edycji, duplikacji, usuwania wydarzenia i usuwania wszystkich wydarzeń.
-dodano powiadomienia pojawiajace sie przy błedzie w przypadku
klikniecia delete all a nie ma wydarzeń w tabelki, gdy wystąpi
błąd przy dodawaniu wydarzenia.
