+++ iowebapp/src/iowebapp/IowebappUI.java
@@ -1,8 +1,5 @@
 package iowebapp;
 
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.Date;
 import java.util.Locale;
 import javax.servlet.annotation.WebServlet;
@@ -13,10 +10,11 @@ import com.vaadin.event.ItemClickEvent;
 import com.vaadin.event.ItemClickEvent.ItemClickListener;
 import com.vaadin.server.FileDownloader;
 import com.vaadin.server.FontAwesome;
+import com.vaadin.server.Page;
 import com.vaadin.server.StreamResource;
-import com.vaadin.server.StreamResource.StreamSource;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.server.VaadinServlet;
+import com.vaadin.shared.Position;
 import com.vaadin.ui.*;
 import com.vaadin.ui.Button.ClickEvent;
 import com.vaadin.ui.Button.ClickListener;
@@ -128,6 +126,10 @@ public class IowebappUI extends UI {
 					 */
 					@Override
 					public void buttonClick(final ClickEvent event) {
+						final Notification editSuccess = new Notification("Successfully deleted event!", Notification.Type.ASSISTIVE_NOTIFICATION);
+						editSuccess.setPosition(Position.TOP_CENTER);
+						editSuccess.setDelayMsec(1000);
+						editSuccess.show(Page.getCurrent());
 						eventsContainer.removeItem(rowReference.getItemId());
 
 					}
@@ -141,6 +143,10 @@ public class IowebappUI extends UI {
 					 */
 					@Override
 					public void buttonClick(final ClickEvent event) {
+						final Notification editSuccess = new Notification("Successfully copied event!", Notification.Type.ASSISTIVE_NOTIFICATION);
+						editSuccess.setPosition(Position.TOP_CENTER);
+						editSuccess.setDelayMsec(1000);
+						editSuccess.show(Page.getCurrent());
 						eventsContainer.addBean(new CalendarEvent(eventItem.getTitle(), eventItem.getDateStart(),
 								new Date(), new Date(), eventItem.getDateEnd(), eventItem.getLocation(),
 								eventItem.getDescription(), eventItem.isAllDay()));
@@ -245,8 +251,14 @@ public class IowebappUI extends UI {
 			 */
 			@Override
 			public void buttonClick(final ClickEvent event) {
-				addWindow(new DeleteAllWindow(eventsContainer));
-
+				final Notification deleteError = new Notification("There're no events!", Notification.Type.ERROR_MESSAGE);
+				deleteError.setPosition(Position.TOP_CENTER);
+				deleteError.setDelayMsec(1000);
+				if (eventsContainer.size() >= 1) {
+					addWindow(new DeleteAllWindow(eventsContainer));
+					return;
+				}
+				deleteError.show(Page.getCurrent());
 			}
 		});
 
@@ -307,88 +319,7 @@ public class IowebappUI extends UI {
 
 	}
 
-	/**
-	 * Klasa która konwertuje dane z tabelki do formatu iCal i tworzy strumieñ
-	 * wejœciowy.
-	 * 
-	 * @author Krzysztof Perchlicki
-	 *
-	 */
-	class IcalGenerator implements StreamSource {
-
-		/**
-		 * Tablica przechowuj¹ca referencje do wszystkich wydarzeñ (CalendarEvent).
-		 */
-		private final BeanItemContainer<CalendarEvent> eventsContainer;
-
-		/**
-		 * Konstruktor przyjmuj¹cy eventsContainer.
-		 * 
-		 * @param eventsContainer
-		 */
-		public IcalGenerator(final BeanItemContainer<CalendarEvent> eventsContainer) {
-			this.eventsContainer = eventsContainer;
-
-		}
-
-		/**
-		 * Metoda która konwertuje dane z tabeli do formatu iCal i tworzy
-		 * strumieñ wejœciowy.
-		 */
-		@Override
-		public InputStream getStream() {
-			final StringBuilder icalFile = new StringBuilder(200);
-			final Date dateStamp = new Date();
-			icalFile.append("BEGIN:VCALENDAR\nPRODID:IOWEBAPP_PROJECT_TEAM");
-			for (int i = 0; i < eventsContainer.size(); i++) {
-				final CalendarEvent event = eventsContainer.getItem(eventsContainer.getIdByIndex(i)).getBean();
-				icalFile.append("\nBEGIN:VEVENT\nDTSTART");
-				icalFile.append(dateConvert(event.getDateStart(), event.isAllDay()));
-				icalFile.append("\nDTEND");
-				icalFile.append(dateConvert(event.getDateEnd(), event.isAllDay()));
-				icalFile.append("\nDTSTAMP");
-				icalFile.append(dateConvert(dateStamp, false));
-				icalFile.append("\nCREATED");
-				icalFile.append(dateConvert(event.getDateCreated(), false));
-				icalFile.append("\nDESCRIPTION:");
-				icalFile.append(event.getDescription());
-				icalFile.append("\nLAST-MODIFIED");
-				icalFile.append(dateConvert(event.getDateModified(), false));
-				icalFile.append("\nLOCATION:");
-				icalFile.append(event.getLocation());
-				icalFile.append("\nSUMMARY:");
-				icalFile.append(event.getTitle());
-				icalFile.append("\nEND:VEVENT");
-
-			}
-			icalFile.append("\nEND:VCALENDAR");
-			return new ByteArrayInputStream(icalFile.toString().getBytes(StandardCharsets.UTF_8));
-		}
-
-		/**
-		 * Konwertuje daty na zgodne z formatem zapisu w plikach iCal.
-		 * 
-		 * @param date
-		 * @param checked
-		 * @return dataIcal
-		 */
-		@SuppressWarnings("deprecation")
-		String dateConvert(final Date date, final boolean checked) {
-			final int year = date.getYear() + 1900;
-			final int month = date.getMonth() + 1;
-			final int day = date.getDate();
-			final int hours = date.getHours();
-			final int minutes = date.getMinutes();
-			final int seconds = date.getSeconds();
-			if (!checked) {
-				return String.format(":%d%02d%02dT%02d%02d%02d", year, month, day, hours, minutes, seconds);
-			}
-			return String.format(";VALUE=DATE:%d%02d%02d", year, month, day);
-		}
-
-	}
-
-	StreamResource createResource(final BeanItemContainer<CalendarEvent> eventsContainer) {
+	private StreamResource createResource(final BeanItemContainer<CalendarEvent> eventsContainer) {
 		return new StreamResource(new IcalGenerator(eventsContainer), "plik.ICS");
 	}
